+++ framework_src/native_library/android/FreKotlinExampleANE/FreKotlin/src/main/java/com/tuarua/frekotlin/FREArray.kt
@@ -20,6 +20,7 @@ package com.tuarua.frekotlin
 
 import com.adobe.fre.FREArray
 import com.adobe.fre.FREObject
+
 /**
  * Creates a FREArray.
  * @param [freObject] the FREObject.
@@ -28,56 +29,32 @@ import com.adobe.fre.FREObject
 fun FREArray(freObject: FREObject): FREArray {
     return freObject as FREArray
 }
+
 /**
  * Creates a FREArray.
  * @param [value] the [IntArray] to convert.
  * @return a new FREArray.
  */
 fun FREArray(value: IntArray): FREArray? {
-    return try {
-        val rv = FREArray("int") ?: return null
-        for (i in value.indices) {
-            rv[i] = value[i].toFREObject()
-        }
-        rv
-    } catch (e: Exception) {
-        FreKotlinLogger.log("cannot create FREArray from $value", e)
-        null
-    }
+    return FREArray("int", items = value.map { it.toFREObject() })
 }
+
 /**
  * Creates a FREArray.
  * @param [value] the [LongArray] to convert.
  * @return a new FREArray.
  */
 fun FREArray(value: LongArray): FREArray? {
-    return try {
-        val rv = FREArray("Number") ?: return null
-        for (i in value.indices) {
-            rv[i] = value[i].toFREObject()
-        }
-        rv
-    } catch (e: Exception) {
-        FreKotlinLogger.log("cannot create FREArray from $value", e)
-        null
-    }
+    return FREArray("Number", items = value.map { it.toFREObject() })
 }
+
 /**
  * Creates a FREArray.
  * @param [value] the [DoubleArray] to convert.
  * @return a new FREArray.
  */
 fun FREArray(value: DoubleArray): FREArray? {
-    return try {
-        val rv = FREArray("Number") ?: return null
-        for (i in value.indices) {
-            rv[i] = value[i].toFREObject()
-        }
-        rv
-    } catch (e: Exception) {
-        FreKotlinLogger.log("cannot create FREArray from $value", e)
-        null
-    }
+    return FREArray("Number", items = value.map { it.toFREObject() })
 }
 
 /**
@@ -86,16 +63,7 @@ fun FREArray(value: DoubleArray): FREArray? {
  * @return a new FREArray.
  */
 fun FREArray(value: FloatArray): FREArray? {
-    return try {
-        val rv = FREArray("Number") ?: return null
-        for (i in value.indices) {
-            rv[i] = value[i].toFREObject()
-        }
-        rv
-    } catch (e: Exception) {
-        FreKotlinLogger.log("cannot create FREArray from $value", e)
-        null
-    }
+    return FREArray("Number", items = value.map { it.toFREObject() })
 }
 
 /**
@@ -104,16 +72,7 @@ fun FREArray(value: FloatArray): FREArray? {
  * @return a new FREArray.
  */
 fun FREArray(value: ShortArray): FREArray? {
-    return try {
-        val rv = FREArray("int") ?: return null
-        for (i in value.indices) {
-            rv[i] = value[i].toFREObject()
-        }
-        rv
-    } catch (e: Exception) {
-        FreKotlinLogger.log("cannot create FREArray from $value", e)
-        null
-    }
+    return FREArray("int", items = value.map { it.toFREObject() })
 }
 
 /**
@@ -122,16 +81,7 @@ fun FREArray(value: ShortArray): FREArray? {
  * @return a new FREArray.
  */
 fun FREArray(value: BooleanArray): FREArray? {
-    return try {
-        val rv = FREArray("Boolean") ?: return null
-        for (i in value.indices) {
-            rv[i] = value[i].toFREObject()
-        }
-        rv
-    } catch (e: Exception) {
-        FreKotlinLogger.log("cannot create FREArray from $value", e)
-        null
-    }
+    return FREArray("Boolean", items = value.map { it.toFREObject() })
 }
 
 /**
@@ -140,32 +90,30 @@ fun FREArray(value: BooleanArray): FREArray? {
  * @return a new FREArray.
  */
 fun FREArray(value: List<String>): FREArray? {
-    return try {
-        val rv = FREArray("String") ?: return null
-        for (i in value.indices) {
-            rv[i] = value[i].toFREObject()
-        }
-        rv
-    } catch (e: Exception) {
-        FreKotlinLogger.log("cannot create FREArray from $value", e)
-        null
-    }
+    return FREArray("String", items = value.map { it.toFREObject() })
 }
 
 /**
  * Creates a FREArray.
  * @param [className] name of AS3 class to create.
  * @param [length] number of elements in the array.
- * @param [fixed] whether the array is fixed
+ * @param [fixed] whether the array is fixed.
+ * @param [items] populates the FREArray with the supplied array of FREObjects.
  * @return a new FREArray.
  */
-fun FREArray(className: String, length: Int = 0, fixed: Boolean = false): FREArray? {
-    return try {
-        FREArray.newArray(className, length, fixed)
+fun FREArray(className: String, length: Int = 0, fixed: Boolean = false,
+             items: List<FREObject?>? = null): FREArray? {
+    try {
+        val ret = FREArray.newArray(className, length, fixed)
+        if (items == null) return ret
+        for ((i, v) in items.withIndex()) {
+            if (length > 0) ret[i] = v else ret.push(v)
+        }
+        return ret
     } catch (e: Exception) {
-        FreKotlinLogger.log("cannot create FREArray of $className", e)
-        null
+        FreKotlinLogger.error("cannot create FREArray of $className", e)
     }
+    return null
 }
 
 /**
@@ -178,7 +126,7 @@ operator fun FREArray.set(index: Int, value: FREObject?) {
     try {
         this.setObjectAt(index.toLong(), value)
     } catch (e: Exception) {
-        FreKotlinLogger.log("cannot set FREArray at index $index to value ${value.toStr()}", e)
+        FreKotlinLogger.error("cannot set FREArray at index $index to value ${value.toStr()}", e)
     }
 }
 
@@ -187,7 +135,7 @@ operator fun FREArray.set(index: Int, value: FREObject?) {
  * @receiver The FREArray.
  * @property args One or more values to append to the [FREArray].
  */
-fun FREArray.push(vararg args: Any?) : Long {
+fun FREArray.push(vararg args: Any?): Long {
     return Long(FreKotlinHelper.callMethod(this, "push", args)) ?: 0
 }
 
@@ -209,7 +157,7 @@ fun FREArray.insert(value: FREObject, at: Int) {
  * (for example, -1 for the last element of the Vector).
  * @return the element that was removed from the original FREArray
  */
-fun FREArray.remove(at: Int) : FREObject? {
+fun FREArray.remove(at: Int): FREObject? {
     return this.call("removeAt", at)
 }
 
@@ -227,7 +175,7 @@ operator fun FREArray.get(index: Int): FREObject? {
     return try {
         this.getObjectAt(index.toLong())
     } catch (e: Exception) {
-        FreKotlinLogger.log("cannot get FREArray at index $index", e)
+        FreKotlinLogger.error("cannot get FREArray at index $index", e)
         null
     }
 }
@@ -241,11 +189,6 @@ operator fun FREArray?.iterator(): Iterator<FREObject?> {
     }.iterator()
 }
 
-@Deprecated("Use toFREObject instead", ReplaceWith("IntArray.toFREObject()"), DeprecationLevel.ERROR)
-fun IntArray.toFREArray(): FREArray? {
-    return FREArray(this)
-}
-
 /**
  * Converts a IntArray to a FREArray.
  * @receiver The [IntArray].
@@ -255,22 +198,12 @@ fun IntArray.toFREObject(): FREArray? {
     return FREArray(this)
 }
 
-@Deprecated("Use toFREObject instead", ReplaceWith("BooleanArray.toFREObject()"), DeprecationLevel.ERROR)
-fun BooleanArray.toFREArray(): FREArray? {
-    return FREArray(this)
-}
-
 /**
  * Converts a [BooleanArray] to a FREArray.
  * @receiver The [BooleanArray].
  * @return A new FREArray.
  */
 fun BooleanArray.toFREObject(): FREArray? {
-    return FREArray(this)
-}
-
-@Deprecated("Use toFREObject instead", ReplaceWith("DoubleArray.toFREObject()"), DeprecationLevel.ERROR)
-fun DoubleArray.toFREArray(): FREArray? {
     return FREArray(this)
 }
 
@@ -298,11 +231,6 @@ fun ShortArray.toFREObject(): FREArray? {
     return FREArray(this)
 }
 
-@Deprecated("Use toFREObject instead", ReplaceWith("List<String>.toFREObject()"), DeprecationLevel.ERROR)
-fun List<String>.toFREArray(): FREArray? {
-    return FREArray(this)
-}
-
 /**
  * Converts a List<String> to a FREArray.
  * @receiver The List<String>.
@@ -311,19 +239,22 @@ fun List<String>.toFREObject(): FREArray? {
     return FREArray(this)
 }
 
+/**
+ * Returns a list containing the results of applying the given [transform] function
+ * to each element in the original array.
+ */
+fun <R> FREArray.map(transform: (FREObject?) -> R): List<R> {
+    val list: MutableList<R> = mutableListOf()
+    for (fre in this) {
+        list.add(transform(fre))
+    }
+    return list.toList()
+}
+
 /** Converts a FREArray to a [BooleanArray]. */
 fun BooleanArray(freArray: FREArray?): BooleanArray {
     if (freArray == null) return booleanArrayOf()
-    val count = freArray.length.toInt()
-    val ret = BooleanArray(count)
-    for (i in 0 until count) {
-        val v = Boolean(freArray.getObjectAt(i.toLong()))
-        when {
-            v != null -> ret[i] = v
-            else -> return booleanArrayOf()
-        }
-    }
-    return ret
+    return freArray.map { Boolean(it) }.filterNotNull().toBooleanArray()
 }
 
 /** Converts a FREObject to a [BooleanArray]. */
@@ -335,16 +266,7 @@ fun BooleanArray(freObject: FREObject?): BooleanArray {
 /** Converts a FREArray to a [DoubleArray]. */
 fun DoubleArray(freArray: FREArray?): DoubleArray {
     if (freArray == null) return doubleArrayOf()
-    val count = freArray.length.toInt()
-    val ret = DoubleArray(count)
-    for (i in 0 until count) {
-        val v = Double(freArray.getObjectAt(i.toLong()))
-        when {
-            v != null -> ret[i] = v
-            else -> return doubleArrayOf()
-        }
-    }
-    return ret
+    return freArray.map { Double(it) }.filterNotNull().toDoubleArray()
 }
 
 /** Converts a FREArray to a [DoubleArray]. */
@@ -356,16 +278,7 @@ fun DoubleArray(freObject: FREObject?): DoubleArray {
 /** Converts a FREArray to an [IntArray]. */
 fun IntArray(freArray: FREArray?): IntArray {
     if (freArray == null) return intArrayOf()
-    val count = freArray.length.toInt()
-    val ret = IntArray(count)
-    for (i in 0 until count) {
-        val v = Int(freArray.getObjectAt(i.toLong()))
-        when {
-            v != null -> ret[i] = v
-            else -> return intArrayOf()
-        }
-    }
-    return ret
+    return freArray.map { Int(it) }.filterNotNull().toIntArray()
 }
 
 /** Converts a FREObject to an [IntArray]. */
@@ -377,16 +290,7 @@ fun IntArray(freObject: FREObject?): IntArray {
 /** Converts a FREArray to an [ShortArray]. */
 fun ShortArray(freArray: FREArray?): ShortArray {
     if (freArray == null) return shortArrayOf()
-    val count = freArray.length.toInt()
-    val ret = ShortArray(count)
-    for (i in 0 until count) {
-        val v = Short(freArray.getObjectAt(i.toLong()))
-        when {
-            v != null -> ret[i] = v
-            else -> return shortArrayOf()
-        }
-    }
-    return ret
+    return freArray.map { Short(it) }.filterNotNull().toShortArray()
 }
 
 /** Converts a FREObject to an [ShortArray]. */
@@ -398,16 +302,7 @@ fun ShortArray(freObject: FREObject?): ShortArray {
 /** Converts a FREArray to an [FloatArray]. */
 fun FloatArray(freArray: FREArray?): FloatArray {
     if (freArray == null) return floatArrayOf()
-    val count = freArray.length.toInt()
-    val ret = FloatArray(count)
-    for (i in 0 until count) {
-        val v = Float(freArray.getObjectAt(i.toLong()))
-        when {
-            v != null -> ret[i] = v
-            else -> return floatArrayOf()
-        }
-    }
-    return ret
+    return freArray.map { Float(it) }.filterNotNull().toFloatArray()
 }
 
 /** Converts a FREObject to an [FloatArray]. */
@@ -419,16 +314,7 @@ fun FloatArray(freObject: FREObject?): FloatArray {
 /** Converts a FREArray to a [LongArray]. */
 fun LongArray(freArray: FREArray?): LongArray {
     if (freArray == null) return longArrayOf()
-    val count = freArray.length.toInt()
-    val ret = LongArray(count)
-    for (i in 0 until count) {
-        val v = Long(freArray.getObjectAt(i.toLong()))
-        when {
-            v != null -> ret[i] = v
-            else -> return longArrayOf()
-        }
-    }
-    return ret
+    return freArray.map { Long(it) }.filterNotNull().toLongArray()
 }
 
 /** Converts a FREObject to an [LongArray]. */