+++ framework_src/native_library/android/FreKotlinExampleANE/FreKotlin/src/main/java/com/tuarua/frekotlin/FlashRuntimeExtensions.kt
@@ -158,63 +158,11 @@ fun FREObject(className: String, vararg args: Any?): FREObject? {
     return try {
         FREObject.newObject(className, argsArr)
     } catch (e: Exception) {
-        FreKotlinLogger.log("cannot create new class $className", e)
+        FreKotlinLogger.error("cannot create new class $className", e)
         null
     }
 }
 
-@Deprecated("Use accessor  instead", ReplaceWith("FREObject.set(name: String)"), DeprecationLevel.ERROR)
-fun FREObject.setProp(name: String, value: Any?) {
-    if (value is FREObject) {
-        FreKotlinHelper.setProperty(this, name, value)
-        return
-    }
-    if (value is FreObjectKotlin) {
-        FreKotlinHelper.setProperty(this, name, value.rawValue)
-        return
-    }
-    if (value is String) {
-        FreKotlinHelper.setProperty(this, name, value.toFREObject())
-        return
-    }
-    if (value is Int) {
-        FreKotlinHelper.setProperty(this, name, value.toFREObject())
-        return
-    }
-    if (value is Double) {
-        FreKotlinHelper.setProperty(this, name, value.toFREObject())
-        return
-    }
-    if (value is Float) {
-        FreKotlinHelper.setProperty(this, name, value.toFREObject())
-        return
-    }
-    if (value is Long) {
-        FreKotlinHelper.setProperty(this, name, value.toFREObject())
-        return
-    }
-    if (value is Short) {
-        FreKotlinHelper.setProperty(this, name, value.toFREObject())
-        return
-    }
-    if (value is Boolean) {
-        FreKotlinHelper.setProperty(this, name, value.toFREObject())
-        return
-    }
-    if (value is Date) {
-        FreKotlinHelper.setProperty(this, name, value.toFREObject())
-        return
-    }
-    if (value is Any) {
-        FreKotlinLogger.log("FREObject.setProp - no automatic conversion type found")
-        return
-    }
-}
-
-@Deprecated("Use accessor  instead", ReplaceWith("FREObject.get(name: String)"), DeprecationLevel.ERROR)
-fun FREObject.getProp(name: String): FREObject? {
-    return FreKotlinHelper.getProperty(this, name)
-}
 
 /** Gets the named property of a FREObject */
 operator fun FREObject?.get(name: String): FREObject? {
@@ -254,7 +202,7 @@ fun Int.toFREObject(): FREObject? {
     return try {
         FREObject.newObject(this)
     } catch (e: Exception) {
-        FreKotlinLogger.log("cannot create FREObject from $this", e)
+        FreKotlinLogger.error("cannot create FREObject from $this", e)
         null
     }
 }
@@ -264,7 +212,7 @@ fun Short.toFREObject(): FREObject? {
     return try {
         FREObject.newObject(this.toInt())
     } catch (e: Exception) {
-        FreKotlinLogger.log("cannot create FREObject from $this", e)
+        FreKotlinLogger.error("cannot create FREObject from $this", e)
         null
     }
 }
@@ -274,7 +222,7 @@ fun Boolean.toFREObject(): FREObject? {
     return try {
         FREObject.newObject(this)
     } catch (e: Exception) {
-        FreKotlinLogger.log("cannot create FREObject from $this", e)
+        FreKotlinLogger.error("cannot create FREObject from $this", e)
         null
     }
 }
@@ -284,7 +232,7 @@ fun String.toFREObject(): FREObject? {
     return try {
         FREObject.newObject(this)
     } catch (e: Exception) {
-        FreKotlinLogger.log("cannot create FREObject from $this", e)
+        FreKotlinLogger.error("cannot create FREObject from $this", e)
         null
     }
 }
@@ -294,7 +242,7 @@ fun Double.toFREObject(): FREObject? {
     return try {
         FREObject.newObject(this)
     } catch (e: Exception) {
-        FreKotlinLogger.log("cannot create FREObject from $this", e)
+        FreKotlinLogger.error("cannot create FREObject from $this", e)
         null
     }
 }
@@ -304,7 +252,7 @@ fun Long.toFREObject(): FREObject? {
     return try {
         FREObject.newObject(this.toDouble())
     } catch (e: Exception) {
-        FreKotlinLogger.log("cannot create FREObject from $this", e)
+        FreKotlinLogger.error("cannot create FREObject from $this", e)
         null
     }
 }
@@ -314,7 +262,7 @@ fun Float.toFREObject(): FREObject? {
     return try {
         FREObject.newObject(this.toDouble())
     } catch (e: Exception) {
-        FreKotlinLogger.log("cannot create FREObject from $this", e)
+        FreKotlinLogger.error("cannot create FREObject from $this", e)
         null
     }
 }
@@ -324,24 +272,25 @@ fun Date.toFREObject(): FREObject? {
     return try {
         FREObject("Date", this.time)
     } catch (e: Exception) {
-        FreKotlinLogger.log("cannot create FREObject from $this", e)
+        FreKotlinLogger.error("cannot create FREObject from $this", e)
         null
     }
 }
 
-@Deprecated("Return null instead and set FreKotlinLogger.context = _context", ReplaceWith("null"), DeprecationLevel.ERROR)
-fun FreConversionException(variableName: String): FREObject? {
+/**  returns an ANEError stating which function has not received enough parameters */
+@Deprecated("Use FreArgException() instead", ReplaceWith("FreArgException()"), DeprecationLevel.ERROR)
+fun FreArgException(functionName: String): FREObject? {
     return try {
-        FreException("[FreKotlin] Cannot convert $variableName").getError()
+        FreException("[FreKotlin] Not enough or incorrect arguments passed to $functionName").getError()
     } catch (e: Exception) {
         null
     }
 }
 
 /**  returns an ANEError stating which function has not received enough parameters */
-fun FreArgException(functionName: String): FREObject? {
+fun FreArgException(): FREObject? {
     return try {
-        FreException("[FreKotlin] Not enough or incorrect arguments passed to $functionName").getError()
+        FreException("[FreKotlin] Not enough or incorrect arguments passed to method").getError()
     } catch (e: Exception) {
         null
     }