+++ framework_src/native_library/android/FreKotlinExampleANE/app/src/main/java/com/tuarua/FreKotlinExampleContext.java
@@ -16,56 +16,30 @@
 
 package com.tuarua;
 
-import com.adobe.air.TRActivityResultCallback;
 import com.adobe.air.AndroidActivityWrapper;
-import com.adobe.air.TRStateChangeCallback;
 import com.tuarua.frekotlin.FreKotlinContext;
 import com.tuarua.frekotlin.FreKotlinMainController;
 
 //This file must remain as Java
 @SuppressWarnings({"WeakerAccess", "unused"})
-public class FreKotlinExampleContext extends FreKotlinContext implements TRActivityResultCallback, TRStateChangeCallback {
+public class FreKotlinExampleContext extends FreKotlinContext {
     private AndroidActivityWrapper aaw;
+    private KotlinController kc;
 
     public FreKotlinExampleContext(String name, FreKotlinMainController controller, String[] functions) {
         super(name, controller, functions);
-        this.controller = controller;
+        kc = (KotlinController) this.controller;
         aaw = AndroidActivityWrapper.GetAndroidActivityWrapper();
-        aaw.addActivityResultListener(this);
-        aaw.addActivityStateChangeListner(this);
-    }
-
-    @Override
-    public void onActivityStateChanged(AndroidActivityWrapper.ActivityState activityState) {
-        super.onActivityStateChanged(activityState);
-        switch (activityState){
-            case STARTED:
-                this.controller.onStarted();
-                break;
-            case RESTARTED:
-                this.controller.onRestarted();
-                break;
-            case RESUMED:
-                this.controller.onResumed();
-                break;
-            case PAUSED:
-                this.controller.onPaused();
-                break;
-            case STOPPED:
-                this.controller.onStopped();
-                break;
-            case DESTROYED:
-                this.controller.onDestroyed();
-                break;
-        }
+        aaw.addActivityResultListener(kc);
+        aaw.addActivityStateChangeListner(kc);
     }
 
     @Override
     public void dispose() {
         super.dispose();
         if (aaw != null) {
-            aaw.removeActivityResultListener(this);
-            aaw.removeActivityStateChangeListner(this);
+            aaw.removeActivityResultListener(kc);
+            aaw.removeActivityStateChangeListner(kc);
             aaw = null;
         }
     }