+++ starter_project/example/get_android_dependencies.sh
@@ -1,2 +1,2 @@
 #!/bin/sh
-wget -O ../native_extension/ane/mobile/com.tuarua.frekotlin-1.8.0.ane https://github.com/tuarua/Android-ANE-Dependencies/blob/master/anes/kotlin/com.tuarua.frekotlin-1.8.0.ane?raw=true
+wget -O android_dependencies/com.tuarua.frekotlin-1.9.1.ane https://github.com/tuarua/Android-ANE-Dependencies/blob/master/anes/kotlin/com.tuarua.frekotlin-1.9.1.ane?raw=true